+++ dag/dag_case1.py
@@ -1,5 +1,5 @@
-from dag_structures import Node, PointNode, SegNode, Leaf
-from geometric_structures import Trapezoid
+from data_structures.dag_structures import Node, PointNode, SegNode, Leaf
+from data_structures.geometric_structures import Trapezoid
 
 
 def handle_case1_left(left_endpoint, seg, node_leaf, degenerate_point):
