+++ test_data/6_case.txt
@@ -1,6 +1,8 @@
-4
+6
 0 0 100 100
 40 5 95 15
 30 60 70 60
 60 55 95 15
 5 40 55 20
+20 50 35 90
+3 50 70 20
