+++ test_data/hll7371.txt
@@ -1,0 +1,8 @@
+4
+0 0 100 100
+40 5 95 15
+30 60 70 60 
+60 55 95 15
+5 40 55 20
+
+
